+++ firmware/uc_master/platformio.ini
@@ -12,7 +12,7 @@
 platform = atmelavr
 board = test
 framework = arduino
-upload_port = /dev/cu.wchusbserial1410
+upload_port = /dev/cu.wchusbserial1420
 
 
 #[env:nanoatmega328]
