+++ firmware/uc_slave/lib/slibrary/slibrary.h
@@ -95,12 +95,104 @@ String  mix_var    = "";   String  mix_set    = "";
 String  temp_var   = "";   String  temp_set   = "";
 
 
+//_BV(x) = 1 << x
+inline void set_motor ( uint16_t *count,
+                        uint16_t *count_set,
+                        volatile uint8_t *MOT,
+                                 uint8_t START )
+{
+  if ( *count == *count_set ) {
+    *MOT |= START;
+    *count = 0;
+  }
+
+  else if ( *count == LIMIT ) {
+    *MOT &= ~START;
+  }
+
+  (*count)++;
+  return;
+}
+
+//_BV(x) = 1 << x
+//rstx=0 (enable); dirx=1 (cw), else ccw.
+inline void setup_dir_rst ( uint8_t RST,   uint8_t DIR,  uint8_t *var,
+                            uint8_t *rstx, uint8_t *dirx,
+                            volatile uint8_t *PORT_1,
+                            volatile uint8_t *PORT_2 )
+{ //PORT_1:
+  if( !(*rstx) ) {
+    if ( !(*var) )
+      *PORT_1 &= ~RST;
+    else
+      *PORT_1 |=  RST;
+  }
+  else
+    *PORT_1 &= ~RST;
+
+  //PORT_2:
+  if ( *dirx )
+    *PORT_2 |=  DIR;
+  else
+    *PORT_2 &= ~DIR;
+}
+
+//ISR: Function of Interruption in timer one. _BV(x) = 1 << x
+void motor_control() {
+  //PORT_CONTROL = 1 << START_CONTROL; //PIN UP
+  // set_motor(&count_m1, &count_m1_set, &MOT1, _BV(START1) );
+  // set_motor(&count_m2, &count_m2_set, &MOT2, _BV(START2) );
+  set_motor(&count_m3, &count_m3_set, &MOT3, _BV(START3) );
+  set_motor(&count_m4, &count_m4_set, &MOT4, _BV(START4) );
+  set_motor(&count_m5, &count_m5_set, &MOT5, _BV(START5) );
+  //PORT_CONTROL = 0 << START_CONTROL;   //Pin DOWN
+}
+
+
+inline void time_setup ( uint8_t motor_speed, volatile uint16_t *count_set, uint16_t *count ) {
+  if ( !motor_speed ) {
+    *count = 0;
+    *count_set = 0;
+    return;
+  }
+  else {
+    *count = 0;
+    //Transform of motor_speed [rpm] to time in [us] and SETTING FOR THRESHOLD OF TIME:
+    if      ( CONVERTION(motor_speed) <= CONVERTION(SPEED_MAX) )  *count_set = CONVERTION(SPEED_MAX);
+    else if ( CONVERTION(motor_speed) >= CONVERTION(SPEED_MIN) )  *count_set = CONVERTION(SPEED_MIN);
+    else                                                          *count_set = CONVERTION(motor_speed);
+
+    return;
+  }
+}
+
+
+void serialEvent() {
+  while (Serial.available()) {
+    char inChar = (char)Serial.read();
+    message += inChar;
+    if (inChar == '\n') {
+      stringComplete = true;
+    }
+  }
+}
+
+
+//clean strings
+void clean_strings() {
+  stringComplete = false;
+
+  ph_var     = "";  feed_var = "";  unload_var = "";  temp_var = "";
+  mix_var    = "";  ph_set   = "";  feed_set   = "";  temp_set = "";
+  unload_set = "";  mix_set  = "";  message    = "";
+}
+
+
 //message format write values: wph14.0feed100unload100mix1500temp100rst111111dir111111
 int validate_write() {
 
   if (
     message[0] == 'w'                     &&
-    message.length() == 56                &&
 
     message.substring(1, 3)   == "ph"     &&
     message.substring(7, 11)  == "feed"   &&
@@ -146,14 +238,14 @@ int validate_write() {
     ( message[53] == iINT(1) || message[53] == iINT(0) ) &&
     ( message[54] == iINT(1) || message[54] == iINT(0) )
   )
-  return 1;
+    return 1;
 
   else
     return 0;
 }
 
 
-void crumble() {
+void crumble() {  //se puede alivianar usando .toFloat() directamente despues de substring
   //Serial.println("good");
   ph_var = message.substring(1, 3);
   ph_set = message.substring(3, 7);
@@ -186,98 +278,4 @@ void crumble() {
   dir4 = INT(message[52]);  dir5 = INT(message[53]);  dir6 = INT(message[54]);
 
   return;
-}
-
-
-//_BV(x) = 1 << x
-inline void set_motor ( uint16_t *count,
-                        uint16_t *count_set,
-                        volatile uint8_t *MOT,
-                                 uint8_t START )
-{
-  if ( *count == *count_set ) {
-    *MOT |= START;
-    *count = 0;
-  }
-
-  else if ( *count == LIMIT ) {
-    *MOT &= ~START;
-  }
-
-  (*count)++;
-  return;
-}
-
-//_BV(x) = 1 << x
-//rstx=1 (enable); dirx=1 (cw), else ccw.
-inline void setup_dir_rst ( uint8_t RST,   uint8_t DIR,  uint8_t *var,
-                            uint8_t *rstx, uint8_t *dirx,
-                            volatile uint8_t *PORT_1,
-                            volatile uint8_t *PORT_2 )
-{ //PORT_1:
-  if( !(*rstx) ) {
-    if ( !(*var) )
-      *PORT_1 &= ~RST;
-    else
-      *PORT_1 |=  RST;
-  }
-  else
-    *PORT_1 &= ~RST;
-
-  //PORT_2:
-  if ( *dirx )
-    *PORT_2 |=  DIR;
-  else
-    *PORT_2 &= ~DIR;
-}
-
-//ISR: Function of Interruption in timer one. _BV(x) = 1 << x
-void motor_control() {
-  //PORT_CONTROL = 1 << START_CONTROL; //PIN UP
-  // set_motor(&count_m1, &count_m1_set, &MOT1, _BV(START1) );
-  // set_motor(&count_m2, &count_m2_set, &MOT2, _BV(START2) );
-  set_motor(&count_m3, &count_m3_set, &MOT3, _BV(START3) );
-  set_motor(&count_m4, &count_m4_set, &MOT4, _BV(START4) );
-  set_motor(&count_m5, &count_m5_set, &MOT5, _BV(START5) );
-  //PORT_CONTROL = 0 << START_CONTROL;   //Pin DOWN
-}
-
-
-inline void time_setup ( uint8_t motor_speed, volatile uint16_t *count_set, uint16_t *count ) {
-  if ( !motor_speed ) {
-    *count = 0;
-    *count_set = 0;
-    return;
-  }
-  else {
-    *count = 0;
-    //Transform of motor_speed [rpm] to time in [us]
-    //SETTING FOR THRESHOLD OF TIME
-    if      ( CONVERTION(motor_speed) <= CONVERTION(SPEED_MAX) )  *count_set = CONVERTION(SPEED_MAX);
-    else if ( CONVERTION(motor_speed) >= CONVERTION(SPEED_MIN) )  *count_set = CONVERTION(SPEED_MIN);
-    else                                                          *count_set = CONVERTION(motor_speed);
-
-    return;
-  }
-}
-
-
-void serialEvent() {
-  while (Serial.available()) {
-    char inChar = (char)Serial.read();
-    message += inChar;
-    if (inChar == '\n') {
-      stringComplete = true;
-    }
-  }
-}
-
-
-//clean strings
-void clean_strings() {
-  stringComplete = false;
-
-  ph_var     = "";  feed_var = "";  unload_var = "";  temp_var = "";
-  mix_var    = "";  ph_set   = "";  feed_set   = "";  temp_set = "";
-  unload_set = "";  mix_set  = "";  message    = "";
 }
