+++ firmware/uc_slave/platformio.ini
@@ -12,4 +12,4 @@
 platform = atmelavr
 board = test
 framework = arduino
-upload_port = /dev/cu.wchusbserial1410
+upload_port = /dev/cu.wchusbserial1420
